+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Autonomous/LeftAuto.java
@@ -11,9 +11,10 @@ import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 
 import org.firstinspires.ftc.teamcode.Autonomous.AutonomousIndex.RamsetteCommand;
-import org.firstinspires.ftc.teamcode.Commands.Baskets.AutoHighBasket;
+import org.firstinspires.ftc.teamcode.Commands.Arm.MoveArm;
+import org.firstinspires.ftc.teamcode.Commands.Elevator.ElevatorPositions;
 import org.firstinspires.ftc.teamcode.Commands.Intake.MoveIntake;
-import org.firstinspires.ftc.teamcode.Commands.StowAll;
+import org.firstinspires.ftc.teamcode.Commands.Wrist.MoveWrist;
 import org.firstinspires.ftc.teamcode.Subsystems.Chassis;
 import org.firstinspires.ftc.teamcode.Subsystems.Constants;
 import org.firstinspires.ftc.teamcode.Subsystems.Elevator;
@@ -25,7 +26,7 @@ import java.util.Arrays;
 
 
 @Autonomous
-public class HighBasketAndPark extends LinearOpMode {
+public class LeftAuto extends LinearOpMode {
 
     @Override
     public void runOpMode() throws InterruptedException {
@@ -47,41 +48,12 @@ public class HighBasketAndPark extends LinearOpMode {
 
         Trajectory First = TrajectoryGenerator.generateTrajectory(Arrays.asList(
                 new Pose2d(0.0,0.0,Rotation2d.fromDegrees(0)),
-                new Pose2d(1.15, -0.23,Rotation2d.fromDegrees(45))), ForwardConfig
+                new Pose2d(1.0, 0.0,Rotation2d.fromDegrees(0))), ForwardConfig
         );
-
-        Trajectory Second = TrajectoryGenerator.generateTrajectory(Arrays.asList(
-                new Pose2d(1.07,-0.15,Rotation2d.fromDegrees(45)),
-                new Pose2d(1.15, -0.23,Rotation2d.fromDegrees(45))), ForwardConfig
-        );
-
-        Trajectory Third = TrajectoryGenerator.generateTrajectory(Arrays.asList(
-                new Pose2d(1.15,-0.23,Rotation2d.fromDegrees(-135)),
-                new Pose2d(1.07, -0.15,Rotation2d.fromDegrees(0))), ReverseConfig
-        );
-
-
 
         SequentialCommandGroup FirstCommandGroup = new SequentialCommandGroup(
 
-                new RamsetteCommand(chassis, First),
-                new AutoHighBasket(arm, elevator, wrist),
-                // new RamsetteCommand(chassis, Second),
-                new MoveIntake(intake, Constants.Intake.INTAKE_OPEN).withTimeout(1500),
-                new StowAll(arm, elevator, wrist)
-                /*new RamsetteCommand(chassis, Third),
-                new StowAll(arm, elevator, wrist),
-                new RamsetteCommand(chassis, Third),
-                new GroundGrabMedium(arm, elevator, wrist),
-                new MoveIntake(intake, Constants.Intake.INTAKE_STOW),
-                new StowAll(arm, elevator, wrist),
-                new TurnToAngle(chassis, Rotation2d.fromDegrees(48)),
-                new WaitCommand(1000),
-                new HighBasket(arm, elevator, wrist),
-                new MoveIntake(intake, Constants.Intake.INTAKE_OPEN),//Pending Test
-                new StowAll(arm, elevator, wrist),
-                new MoveIntake(intake, Constants.Intake.INTAKE_STOW)*/
-
+                new RamsetteCommand(chassis, First)
         );
 
 
@@ -104,3 +76,6 @@ public class HighBasketAndPark extends LinearOpMode {
         }
     }
 }
+
+
+
